So, I did give the idea a go mentioned last commit, of only ever using the global Renderer pointer. It's OK? I guess the trouble with this style is that you can't tell when things are overly coupled. The temptation is to hook into some global variable rather than devise a better system. An unexpected upside is that now there is only one variable in the entire program called "renderer". So the fact it's the same one that's accessed everywhere is clearer.
So I suppose what it comes down to is discipline. Discipline to only access things that I should access. Which is already how I try to program so it's no big deal? And also discipline to only make things global when there is guaranteed only one of them, or else keep their associated functions taking a parameter so that you specify which one you're dealing with.

CLOC: 13039