Still in the middle of this... I managed to find the culprit: glTexImage2D() (and probably related functions) hates data that's a non-multiple-of-4-bytes wide, and pads it. That's why I was getting 3 tiles skipped per row - the city width at the time was 133, which %4 gives 1, so it was skipping the next 3 bytes. I get that this would be helpful in situations where you're declaring the array as 2D, but I'm *not*, so yeah...
Not yet sure how to go about rectifying it. I currently have 2 cases that work, because I've made them send the whole city's data as the texture and the city is now 128x128, but the other cases still send the visibleTileBounds so they're all kinds of busted. I probably do want to send the whole city's data, and keep it cached on the GPU, but I've had enough of this for today!