Accidentally doing multiple barely-related changes. /fp
Introduced the GameClock struct, and then realised that it should exist
in the Game not the City, and our loading code only touches the City, so
now I'm correcting that to try and make other parts of Game be saveable.
(The RNG stuff needs saving for one thing.)

Current progress: beginNewGame() is now different from loadExistingGame().
The latter is incomplete and the loading happens elsewhere, but it's a
start.

CLOC: 20780