Switch loadSaveFile() to take the GameState directly
Got a bit muddled, then decided it makes most sense that loading a game
creates a blank game state, then fills it in. So we have newGameState()
instead of loadExistingGame(). beginNewGame() is now looking pretty
redundant, but I'll keep it for now. We'd like to change that flow to
put the game in a "design the map" mode when first creating a new city.

CLOC: 20777