Correct game-clock saving and loading
I originally planned to store the year, month and day so that changes to
our "internal-clock <-> displayed-date" calculations don't cause the
date to jump when loading a previous save. I then completely forgot
about that and stored the currentDay number instead! (Oops.) So, we're
actually storing the year, month and day now.

CLOC: 20993