A conundrum!
We want entities and other objects to be able to record a game date for
when something happened. eg, when a fire started, or when a building was
built or lost power. The trouble is, how do we save that? Right now, we
save the current date as y/m/d, to avoid it jumping if we later change
the date calculation. That's fine, except we'd also have to save ALL
timestamps as y/m/d too or else those would go out of sync, which is a
LOT worse than just the date jumping. So, options:

1) We only save timestamps, for everything. We might get a discontinuity
at some point, but everything in the game will run off timestamps, and
they are quick and simple.
2) We only save y/m/d, for everything. This prevents discontinuities,
but requires more effort to encode and decode them. (It's also 2 extra
bytes per date, but that's pretty negligible.)

I'm not awake enough to make the decision today, so I'll let it stew
for a bit.

CLOC: 21007