Amend GameState initialisation
My intention was to set up a new GameState and make sure it's valid (eg,
that the savedgame loaded correctly) and then make it the global
GameState. There are multiple problems with this:

1) It means having two GameStates in memory at once, which is going to
   be rather hefty. (And previously I was leaking the old one!)
2) More relevant, it means we can't use that global GameState at any
   point while initialising a new City or loading a save. And we do want
   to use it! To make that bug free, we'd need to basically eliminate
   using the global at all, in which case it might as well not exist.

Getting rid of the global might be a good idea, but having to pass the
GameState into so many places is a tangly mess too. IDK.

So for now, I've gone back to only having one GameState. When we create
a new one, we first free the old one. We then immediately assign it to
the global, so it is safe to use.

CLOC: 21029