Swap to simple timestamp saving
The amount of processing needed to convert the simulation's timestamps
to/from a Y/M/D format for the save files just seems really unnecessary.
If all we want is the timestamp, why not just save the timestamp?

So I am!

It's a whole lot simpler. I've also added saving and loading of creation
dates for buildings and fires.

CLOC: 21016