Split save-file code into layers
endian.h has the little-endian conversion code
binary_file.* has code for basic binary-file data types and structures
save_file.* has structs and code specific to the save-file format

Currently it's not a clean split - we're validating always that the file
header says CITY and has the SAV version number. But trying to tease
that out is complicated so I'll do that next.

CLOC: 21033