Move save/load window closer to my design for it
The design is totally ripping-off Factorio's, but theirs is good, so
why not!

The UI system is lacking a few pieces that would make the design work.
I've added them to the todo list.

What we want is a proper way to dynamically lay things out. Like:
- Start at the bottom of the window and add the buttons
- Then, take the rest of the space and divide it into two columns

But right now that's tricky. Part of the problem is that the save/load
window code relies on the saves list being output before the buttons
are. Can probably untangle that but I'm not mentally up to it today.

CLOC: 21080