Fix console shortcuts running repeatedly if key is held down
wasShortcutJustPressed() was just looking at whether the key is NOW down,
so fixed that, and then it was also responding to repeat-keys when held
down, so I added a distinction between real key-presses and repeat ones
from holding the key down, as the "ignoreREpeats" parameter to
keyJustPressed(). I'm sure it'll come in handy elsewhere.

CLOC: 21121