Factor out duplicate UITheme code
Rather than the incredibly repetitive chain of if-elses with duplicate
logic for reading in style properties, we now define the properties and
where they apply in a table in initUiTheme().

After a lot of failed attempts, this new strategy works. Rather than
trying to make the system recognise different style structs, we now only
write to a generic UIStyle struct, which contains every property. Then,
at the end, we copy the data from that struct into the specific one, eg
UIButtonStyle.

The console's outputTextColor property is still a special case, and we
have room for other special-case properties like it, but the majority of
things will go nicely in the hash table.

CLOC: 21070