Cache style lookups with UIStyleReference
Similar to FontReference, except with type checking for the different
types of styles. (Copied from getEntityData())

It's a little hairy... I couldn't see a way to squeeze the type in there
without creating a PropType for each kind of style. So, we put the type
in the constructor for UIStyleReference, and then try hard not to mess
it up. That means making sure we actually call the constructor, which is
not something we do in this codebase otherwise so that's a little
sketchy, but I patched up the two places it was having problems. If we
do mess it up, the game will immediately crash so it's hard to miss, at
least!

CLOC: 21242