Correct previous commit
I didn't like that the function both checked and retrieved, and now that
we have the lambdas we can easily put that code in findStyle(). While
doing so, I realised the previous commit's code fetched the style EVERY
TIME, which makes the whole cache pointless, so that's now rectified.

CLOC: 21143