Fix zooming weirdness, and re-enable keyboard shortcuts
Basically, I'd messed up the logic slightly when converting the snapping
into a function, so the zoomDelta was 10x as powerful.

I'm not sure this is the zoom system we want for the final game - a set
of fixed zoom levels you jump between might be better? But this is an
improvement anyway.

CLOC: 21154