Start untangling popup menu code
uiMenuButton() is basically a bad idea. We're moving to two steps: one
checks the button was pressed and shows/hides the menu; the other sees
if the menu is visible and then renders it. The menu button is a little
repetitive so maybe uiMenuButton() will be reimplemented but it won't
return whether to show the menu as it now does.

CLOC: 21715