Fix bug in the panel scissor behaviour
It's still a little quirky. A Panel creates its background and scissor
the first time it has a widget (or sub-panel) added. This means you CAN
create a row or column to divide up space, but then wait until later to
fill it, which is the case where the bug appeared. But you can't
alternate between adding widgets to one panel then another like A B A...
because the scissor will match whichever panel was last to have its
first widget added.

Hard to explain! But it seems unlikely that I'll ever want to be adding
to two different panels at the same time, so it's fine for now.

CLOC: 21743