Add scrollbars to panels
Only vertical scrollbars are actually tested, though I've implemented
horizontal ones too. It's likely there are typos and bugs there, but as
I don't have anything to test horizontal scrolling with, I'm leaving it.
I think it's better to put the code in now while I'm thinking of it,
rather than leave it (and probably put off adding it when it IS needed.)

CLOC: 21808