Add SpriteRef, as a replacement of direct Sprite pointers
The Sprite pointer in Entity was crashing us when the assets reloaded,
because the pointer is no longer valid. So, a SpriteRef system
instead!

SpriteRefs keep a Sprite pointer, and track when it was fetched, so we
know in getSprite(SpriteRef*) whether the assets system (which now
tracks the reload time) has been reloaded since then, and can thus
fetch it fresh! This feels like an elegant way of handling it.

Also cleared out drawBuildings() which was no longer used at all! Must
have missed that when working on Entities before.

We're still using the old version of getSprite() in a few places.
Probably want to migrate to SpriteRefs for some of them, it looks like
we're keeping SpriteGroup pointers around which is the same issue.
Maybe a generic AssetRef would work for that? Something to consider.

CLOC: 21767