Start replacing the data view UI with a Panel
This is less straightforward than it seems. Our gradient indicator
requires two labels, at the top and bottom. That requires some
sub-panels to divide up that space. Discovered while doing that, that
row() don't work in bottom-up layouts! I think what needs to happen is,
using the currentLeft... fields instead of the contentArea when
plotting where a row/column will go?

Alternatively, I could hack it to not use sub-panels for the gradient,
but that's unsatisfying.

CLOC: 21800