Convert About window to UIPanel
Decided that converting the windows one at a time would be more
straightforward than trying to convert the window_X() functions one at
a time. (Which I'm not sure would be possible at all!) So, a lot of
window code will go after I'm done with all this. For now it's messy.

CLOC: 21889