Fix scissor
Multiplying by the camera matrix was a mistake. Firstly because the
matrix could easily not return an axis-aligned rectangle; and secondly
because it introduces tiny rounding errors which meant the scissor was
sometimes one pixel off. This was leaving gaps between window title
bars and their content panel. So now, scissors are in window-pixel
space, and we just flip the Y coordinate.

CLOC: 21416