Start on UIBackground, as a replacement for background colours
The idea is, you tell it to prepare the placeholder, then you later tell
it to fill it, and it figures out what the placeholder should be
depending on what kind of background it is - without the user code
needing to know or care.

Right now, it's only used for Panels, but I'll start converting other
things next.

CLOC: 21568