Start implementing Sprites as UIDrawables
We read them in, but don't render them. Need to implement Sprite
placeholders in the renderer, where we currently only have regular
rects. (Could maybe make DrawRectPlaceholder its own struct, with a
pointer to the setTexture command struct if it exists, so we can write
to that.)

A couple of necessary tangents have made this more complicated than I'd
hoped. First off, we had awkwardness with SpriteRef not being declared
before the theme, so now we have a separate asset.h file with just the
asset struct definitions. (The ideal solution would be to move the
Assets.theme field to being a proper asset, or a separate entity
entirely, but for now, splitting out a header file does the job and is
much easier.)

CLOC: 21677