Make SpriteDefs actually record their child assets
Went to add ninepatch support to it, realised we were never even
recording the sprites!

Probably want to combine a few asset-def file formats together. Cursors,
palettes and sprites could all be combined at least. Ideally, we'd put
all of the data things in a single file format but IDK. Some of those
declare assets and some don't. (eg, a BuildingDef is not an asset)

CLOC: 21976