Fix building variants showing placeholder sprite
We weren't interning the sprite name for variants. It must have worked
accidentally at some point, but we've worked on the asset system a lot
lately.

As a bonus, we now check strings are valid when attempting to intern
them, so we should catch this earlier in the future.

CLOC: 22008