Remove UITheme.fontNamesToAssetNames
The mapping is now done inside loadUITheme(), which is more
straightforward. The only extra limitation is now fonts need to be
declared before they are used, but that's no big deal. I also made the
font system more robust, so that the placeholder font asset doesn't
crash the game when it's used.

CLOC: 21998