Convert UITheme to an asset
Rather than a set of HashTables, it's now a set of packed Arrays, all
stored in the Asset's data Blob. This does mean we have to do a linear
search through those arrays when getting a style, but as those arrays
are all small, it should be fine. (And we can always replace them with
a fixed-memory HashTable later if needed.)

CLOC: 22056