Replace FontReference uses with AssetRef
The only place that getFont(String) is used, is for displaying the debug data in renderDebugData(). I experimented with putting an AssetRef for the font in DebugState, but since the debug system is initialised before the asset system, we can't actually fetch an AssetRef! (We can't intern the asset name String.)

CLOC: 22055