Implement a deltaTime multiplier for debugging
The deltaTime now lives in globalAppState, replacing the few remaining
uses of the SECONDS_PER_FRAME constant, so everything agrees on how
much time has passed this frame now. Then, as I could, I added a speed
multiplier, which can be read/written using the `speed` command. This
is helpful for testing animations, as we can slow the game right down
and watch if things are happening correctly.

Moved the AppState into its own file, because it's starting to get more
involved than just a plain struct, and it felt awkward it living
half-way down the #include list in main.cpp.

CLOC: 22484