Implement a confirmation dialog for deleting saved games
Also: Fixed that buttons weren't correctly reporting when clicked, but
would repeatedly report the click if you held the mouse down.

Originally thought we'd have a single confirmation dialog that we call
with a custom text, but in cases where that text is generated, that
requires having somewhere to save that text to, in a way that won't
leak memory but still lives as long as the dialog does. So, a
confirmation dialog WindowProc per use case is the way for now.

One possible option for the memory, is to have a preallocated string of
a large size and copy the data into it, as with Toasts. But then does
that live in the UIState, or in the caller somehow? If the UIState, how
do we handle the possibility of multiple dialogs at once? (Which should
never happen. So maybe that's not a real problem.)

CLOC: 22509