Replace findXStyle() functions with findStyle<Type>()
The main motivation was the new findStyle(styleName, defaultStyle)
function, which replaces the repetitive logic we were using when a
styleName is passed in places. But generally, being able to use a
template for things like that is handy.

It's actually less code, except we had to predeclare the findStyle
() templates and then define them in a different file, because of
annoying dependency order stuff.

CLOC: 22514