Close the pause menu when a game is loaded
This is a TEMPORARY solution. The idea of modifying the openWindows
array is flawed, because it is likely that said array is being iterated
at the time. (And in fact, it is being iterated when we load a game, as
that's called from inside the save/load window!) So, a more robust
solution is my next task.

CLOC: 22532