Implement robust window-closing
We now keep a set of all window indices that have been requested to
close. In the updateWindows() loop, we skip any that are requested to
close, and then afterwards, we remove them all in order from highest
index to lowest, so that the indices don't change under us. Also, we
ignore the newActiveWindow if it was in that set.

Should be pretty solid now! Even if we request windows to close from
within a WindowProc, it won't mess things up.

CLOC: 22577