Implement click-and-drag of scrollbar thumb
Before, we just jumped the thumb to wherever the mouse was. We still do
that if the click started off of the thumb, but might remove that
later.

Renamed 'knob' to 'thumb', which is the more standard name.

We have a getScrollbarThumbBounds() function now, but don't use it for
drawing the scrollbar because drawScrollbar() still doesn't take a
ScrollbarState yet. Must get on that.

We also have an initScrollbar() function finally, because they now have
the concept of whether they're horizontal or vertical. A lot of the
code still assumes vertical, but in new stuff I'm writing I'm trying to
handle both as I go along, so there's less to change when I do properly
implement horizontal scrollbars.

CLOC: 2641