Fix scrollbar bugs
Both were the same cause - we were comparing the distance the thumb was
dragged, with the size of the content overflow, not the size of the
scrollbar. It was just a coincidence that the two were almost
identical, so it was a minor difference! The difference with a small
console was much more pronounced.

CLOC: 22640