Make scrollbar mouse-wheel step size configurable
This is a little weird. I've added it as a parameter to initScrollbar(),
but nobody actually provides it there. The console can't, because when
it gets initialised, the asset system isn't loaded yet so it can't
measure its font's line-height. So, it just stuffs that value manually
before each call to updateScrollbar().

We could add it as a parameter to updateScrollbar(), but Panel doesn't
expose that to user code, so Panels wouldn't be able to use it at all.
So, manually writing to the ScrollbarState.mouseWheelStepSize is how
we're doing things.

CLOC: 22637