Test horizontal scrollbars
So far, they do work. Most of the issues have been with Panels not
handling them correctly. There's a bug with windows (or I guess
probably any height-shrink panel?) cutting off the scrollbar with their
scissor. My thoughts are in the TODO list, but I'm not convinced my
assumptions are correct.

CLOC: 22667