Fixed horizontal scrollbar position in panels
I had a long note that I hadn't committed yesterday, about how I thought
the scissor was the problem, and how I'd have to use a placeholder
system for that... and then realised today that the scrollbar was just
being put at the wrong y position. So, problem solved! CLOC: 22668