Hide the scrollbar thumb when the content is too small to need one
I suppose a different approach here would be to draw the thumb as taking
the whole height of the scrollbar track. But that's a style choice. In
any case, this is better than previously, where it would draw a
non-interactive thumb that only took up 2/3 of the scrollbar, which is
really weird and confusing.

CLOC: 22681