Think a lot
Making buttons react (or not) properly when a window blocks them
(whether because it's overlapping or it's modal) is complicated.
Ideally, `uiState->mouseInputHandled` would be perfect, BUT our window
system does update and render as two passes, and mouseInputHandled is
always true by the time we get to the render pass. I've investigated
whether we can combine the two passes, but... basically, no. For one,
we iterate windows front-to-back when updating, but back-to-front for
rendering!

One idea would be to store the window index that caused the mouse input
to be handled, and then see if the current window being rendered is <=
that index. But, that means the underlying Panel system having to know
way too much about the window system. (Unless, Panels have a concept
of "isDisabled" along with doUpdate and doRender, to turn off visual
mouse indicators? IDK)

Possibly, we could store button states somewhere? So that they're set in
the update pass, and just read in the render pass. That would solve
this! But, how do we store button states?

So, yeah.

CLOC: 22680