Clear up old debug data before the start of each frame
If something doesn't get tracked in a given frame, it was keeping the
old data. Now, we zero-out that data so that if the thing isn't
tracked (eg, a temporary RenderBuffer that no longer exists) we don't
get bogus data showing up in the debug overlay, left over from when it
DID exist. Hopefully that makes sense.

CLOC: 22799