Fix crash from windows spawning other windows
Basically, rewrote our "make window active" system to queue things up
instead of acting immediately, like the window-closing system.
Otherwise, we were swapping windows around while iterating them, which
is a bad time. It was actually wrong before, but only showed up because
of the Window::renderBuffer not being set when that happens!

Also, expanded all ChunkedArray remove() functions to take a
keepItemOrder parameter. Possibly that should instead be a field on the
ChunkedArray? Hmmm.

CLOC: 22804