Adjust power-line-over-road variant priorities
The road takes priority now. It's generally better, but there are still
glitches if you do a 9x9 of power lines and roads. Basically, we need
some way to prevent building certain configurations, OR a way for the
variants system to say "this is invalid!" and revert back to a road.

CLOC: 22891