Add _h suffix for hashed strings, for hashed command names
Not sure about actually using a HashTable for the commands. That would
make commands not be in alphabetical order. (Though, we only want that
for the `help` command, which could just do the sorting itself.) But
more than that, we don't have a HashTable that takes a fixed amount of
memory, which we ideally want for this (and for UITheme data). So,
that's what we need really.

CLOC: 22905