Implement fixed-size HashTable
You give it some memory, (or an Arena and it gets some memory) and then
it sticks to within that size. There's 2 ways of initialising it, as
noted in the comments, depending on if you want to do the
memory-allocation yourself or let an Arena do it.

Storing keys is awkward still, as that's a separate allocation - in
fact, each HashTable creates a whole Arena just for storing its keys!
That feels excessive, especially with these fixed-size ones. Ideally
we'd allocate a decent buffer for keys in the initial memory
allocation, but we don't know how big each key will be, so it's
tricky.

CLOC: 22944