Convert UITheme style arrays to fixed-size HashTables
HashTable code is getting a little muddly. Basically, the challenge here
is that we avoid searching through the table twice when attempting to
add something, by expanding the table preemptively. Otherwise, the
process of adding a key to an at-capacity table would be:

1) Scan the table to find the entry
2) Expand the table to make room for the new entry
3) Scan the table again to find the actual location

Thinking about it, that's an incredibly rare situation! The table hardly
ever has to expand. So yeah, we can do it more simply.

CLOC: 22957