Make findStyle() default to the style named "default"
Adding hard-coded styles as a guaranteed-to-exist default is trickier,
because what font do we link to?

I was going to add logging of missing styles, but then realised this is
looking a LOT like the regular Assets system, and so maybe styles
should just be assets? Maybe UITheme doesn't need to exist like this?
IDK.

CLOC: 23058