Diagnosed and fixed the sudden jump in memory size
We'd crossed a threshold where a chunk in the allAssets array was larger
than a memory block in the assetsArena. So, all remaining memory in
that first block was wasted. The block size is 1MB at default, so...
that's where our mysterious 1MB disappeared to!

Fixed by reducing the chunk size. It's still about 10x the current asset
count so it makes no performance difference to do so.

CLOC: 22990