Add sprites for the new water-edge tiles
There's stuff in the all.terrain file that's not implemented by the game
yet, so it'll crash or breakpoint when run.

This includes some in-progress ideas for how to represent the edges in
the .terrain file. But, my current idea for it would require 255
entries, manually entered, so that's problematic. I think we can get
away with a lot fewer - currently thinking 80, but that's still
somewhat redundant - so we'll see.

CLOC: 22994