[INCOMPLETE] Start to rewrite save-file loading with the new system
We're lacking a good way to read arbitrary data from the section. The
terrain and building type-tables are a series of inline strings, so
they're variable-length. That makes it hard to read them predictably.
For now I'm using the section's bytes directly, but it might be a good
idea to modify the tables' format to be consistent, with the character
data at the end.

CLOC: 23521