Move save/load code into Layers
As part of this, I had to stub out the Budget and Education layers.
Budget probably doesn't want to be called a Layer at all, but I don't
yet know what it is. Had the option of deleting it, or just adding this
stub, and chose the latter. But yeah, I shouldn't have put in save code
for things that had zero implementation, oh well.

CLOC: 23536