Reduce memory usage when loading binary files like saves
We store the state of the memory arena when the file is loaded, and then
reset to that state after each section. Initial results indicated it
uses 75% less memory, but I realise that's not the high-water mark, but
rather the memory used by the final section. But it saves something at
least.

CLOC: 23546