Start implementing DropDownLists
Oh boy, am I sick of typing DropDownList. But I couldn't find a better
name for them! I guess DropDown would do, oh well.

Got derailed slightly by a crash in the asset system being unhappy with
zeroed-out AssetRefs.

Everything is implemented, up to UI::putDropDownList(), where the actual
hard stuff will go. The plan is that UIState will maintain a pointer to
an info about the currently-open DDL. As only one will ever be open at
a time, that'll work fine. Just set the pointer when it's activated,
and if that's the current one being dealt with, display the panel. So,
shouldn't be *crazy* complicated.

CLOC: 23863