Add a scrollbar to drop-down lists
ARGH I was stuck for so long on a weird bug, which is from RenderBuffers
keeping state when they shouldn't. I really should get rid of
the "currentShader / currentTexture" stuff, it just makes things messy.
I was thinking a while ago that a less low-level API of things
like "drawText" etc would be better than our current one of
setShader/setTexture and things. Just, here's some text and the shader
and texture to use. That would eliminate these issues, and they're so
hard to track down and fix.

CLOC: 23968