Add a graphic for the drop-down arrow
Not using it yet; not sure actually how that should be defined! The lazy
way would be to add it to the DropDownList style, but it feels more
like a property of the button, that it has an icon at one side or the
other. Though I don't know if any other buttons will want an icon like
that? They might do. I can imagine having a green tick or a red X on a
button associated with a random event choice or something.

So, yeah! Let's give buttons icons.

CLOC: 23968