Implement slider behaviour
It's only horizontal, continuous sliders. I'm not sure if we want to do
vertical, but I expect we'll want discrete ones too.

We have a new drag system! Which I think I want to expand a bit. The
actual drag movement calculation is still done by user code when it
should probably live there instead. We'll then move the scrollbar and
window dragging over to that.

CLOC: 24262