Store UI-related mouse position in one place
Previously, user code had to access renderer->uiCamera a lot, which is
awkward, and resulted in the expensive unproject() call for the mouse
position being called repeatedly. Now, we store the mouse position,
click start position, and window size in the UI:: namespace. Originally
I put them in UIState, but having to access them as
`UI::uiState.whatever` felt unwieldy. So, this is an experiment in not
using UIState. The alternative would be to just make UI be a struct
directly, and maybe that's what it *should* be, but for now it's not.

CLOC: 24261