Fix theme sprite colours
It's not really a bug - sprite drawables expect:
    'spriteName index r g b a'
but ninepatches have the colour straight after the name. That index
value was throwing things off, because I forgot it existed. I'm not
using it, and I doubt we'd need to use it - putting multiple sprites
under one name is useful for the game but not the UI. So, I've removed
that 'index' parameter.

CLOC: 24264