Implement discrete sliders
Turns out to be really simple! Just round the value and then adjust the
thumb's position to match.

I'm now thinking we probably want a text display of the value too, at
least on the settings window. Don't necessarily need to make that a
part of the slider itself.

CLOC: 24319