Implement vertical sliders
They use the same thumb graphic as horizontal sliders. If we want to
actually use them, we'll have to define an extra style definition. Or,
maybe we adjust the style to handle horizontal and vertical thumbs
somehow? I'm not sure, and as we don't have a use for them, I'm not
going to guess. In any case, vertical sliders work just as well as
horizontal ones.

We now also have an Orientation enum now, which I should have added
before, which makes the function calls way clearer than an isHorizontal
or isVertical bool.

CLOC: 24384