Remove support for multiple-value settings
It seemed like a good idea at the time, but the added complexity meant
we were asserting that the settings only had 1 value in most cases. So,
just implementing a V2I setting-type is cleaner and simpler.

Switched `Type` enum to a `SettingType` enum class while I was at it, as
it's really only applicable to settings.

CLOC: 24413