Implement radio button groups in Panels
addRadioButtonGroup() is a nicer API than adding each RadioButton
individually. It's not quite working properly yet - we handle
left-alignment but not right or centre: The button is always aligned as
if it had no label, then the label is stuck to the right of it. To do
this properly we'd need to measure and align the label and button as
one unit. Well, we'd need to measure the largest one out of all the
options, and align based on *that*.

CLOC: 24633