Add padding, alignment and background to LabelStyle
We only actually use the LabelStyle properly in UI::putLabel(), which is
only used for Panel RadioButton groups. So, plenty of places need
changing!

However, we might want to get default style-property values in first,
because most of the label styles don't have these set, and not setting
an alignment property results in crashing.

CLOC: 24654