Default values for style properties
Previously, if you didn't specify a value, it was just zeroed-out.
That's fine for some things, but VERY BAD for others! So, now the
UI::Style struct uses a Maybe<> for each property, so we know when
copying to a specific FooStyle struct whether each property was
provided. We can then provide a sensible default if possible.

For properties pointing at another style, the default is the style
named "default". For alignment, it's generally the top-left. For fonts,
it's the placeholder BitmapFont asset. (Which we currently don't
provide properly, so that's still a crash.) For properties with
qualifiers we use the unqualified property. (eg, backgroundHover uses
background) For a lot of other things, it's still a zeroed-out struct,
because that's fine. (eg, padding or contentPadding or a mouseOffset)

CLOC: 24671