Convert main menu and credits away from UI::drawText()
Ended up replacing the main menu UI code with a Panel. It doesn't let us
set the buttons to all be the same size, which is a shame, but we'll
definitely get to that functionality eventually. Right now, the focus
is on getting rid of drawText() in favour of putLabel(), which handles
the styling correctly.

Also renamed Panel::putText() to ::putLabel(), because it's a Label!

CLOC: 24638