Remove UI::drawText and its remaining uses
We now have no warnings again! :D

The game toolbar looks a bit wonky now, because Labels have styling that
it's not prepared for. (And also because the code for it is wonky.)
But, since it's all going to be replaced eventually anyway, I'm leaving
it as it is for now.

I did notice a bug, that tooltips now crash because they have zero
width. I don't know when that was introduced.

CLOC: 24659