Fix Panel getting a lineWidth of 0 for centre/expand-h aligned widgets
Line width was calculated based on the currentLeft or currentRight, but
with CENTRE or EXPAND alignments, we never set either of those so the
line width was apparently 0. This was causing tooltip windows
(which shrink their width to their content) to resize themselves to 0
width the first run through, and then to crash when the text wouldn't
fit.

We now calculate the line size in completeWidget() instead.

CLOC: 24664