Use calculateWidgetBounds() for RadioButtonGroups
The code there is still a bit weird - we do a lot of manual layout for
the radio buttons and labels - but the group as a whole is laid out
with that function. I'm increasingly thinking we might want the ability
to optionally attach a label to radio buttons and checkboxes, which
would make this code simpler.

CLOC: 24661