Make Window-dragging use the UI drag system
We can't actually use the Window* itself as the ID, because they get
swapped around in the openWindows array, so the address is not
constant. To make up for this, Windows now have an `id` which is
unique, and we pretend that's a pointer to them. As noted in the @Hack
comment, it's not perfect, and could cause future issues. However, for
now we have a very small number of things that can be dragged, and only
windows can't use their memory address so it should be fine for now.

CLOC: 24707