Linux: Implement file-watching with inotify
We make a separate inotify instance for each watcher, because that works
best with our code at the moment: We ask "has this directory had any
changes since last time?", which doesn't work nicely with a single data
source that reports changes in any watcher.

We may eventually want a single poll() per frame which checks inotify
watchers and anything else, but right now that's not necessary.