Util: Make Optional compatible with non-default-constructible types
We make the value a union with an array of bytes, which is what we
initialize when the Optional is empty. Unions are sketchy in C++, so we
manually define a destructor that destroys the value if it exists.

I've also added a Maybe -> Optional constructor, to make gradual
adoption easier.