Util: Replace old Alignment enum with a 2-enum struct
HAlign and VAlign are each a single byte, and not arranged as bitfields.
A lot of places only want either horizontal or vertical alignment, so
having them as separate types makes that harder to mess up.

Even though we could conceivably fit both into a single byte, this is
still smaller than the 32-bit value we used before.