Settings: Make Settings type-safe
This is quite a big change!

SettingsState (with BaseSettingsState) is now aware of its own settings,
and is responsible for loading/saving them, etc.

Those settings are each a Setting object. Instead of SettingDef, which
told you how to find a setting using manual memory offsets, a Setting
is both a definition and its value. Its value is typed, so you can read
it and write to it. It can reset itself, and also knows how to save and
load itself from the config file. This is *much* less spooky than
casting memory offsets into something and then poking at it.

Settings itself is now a simple wrapper around the two SettingsStates.

Ideally I'd like the game to be able to pass in a SettingsState of its
own, but I can't figure out a way to do so that isn't much more awkward
to deal with. So for now, we do have a separate BaseSettingsState with
the majority of the logic, but the SettingsState itself is hard-coded
in Settings.

I had to mangle the radio-button-group code a bit to make it work with
EnumSetting. EnumSetting generally is funky and could use some more
thought, but it works for now.