UI: Lay out data-view window's close button before title
We'd previously hit an assertion in calculateButtonSize() if the title
of the data-view window took up too much space horizontally and didn't
leave room for the close button.

I've also made that window wider which is the actual issue, but the fix
is still needed in case we get text that is longer.