UI: Modernize Toast
It's now a class instead of a zero-initialized struct with free
functions. The main functional difference is that instead of using
copyString() to do dubious data-manipulation, we now keep a
StringBuilder in the Toast, pointed to its internal character buffer,
and use that for copying the message in. It's not the most elegant, but
that's the only mutable String type we have and it works well enough.